requirements.txt
pymongo==4.2
tornado==6.2
python-dotenv==0.20
dnspython==2.2